Fix the sudo request to open root procs
The bug was that in order to scan all the root proc/<pid>/smaps files it
was requested to be sudo at least. It is fixed by igoring the error
return from openfile and just gather the information from valid files to
read.

It might need to investigate more with other folks to check if we are
not missing some information psstop from C version works the same way

Signed-off-by: Victor Rodriguez <victor.rodriguez.bahena@intel.com>